feat(ui): add SortableTable component
