+++ AGENT.md
@@ -46,6 +46,7 @@
 | `FileValidator`              | Vérifie l'extension et la taille           | `apps/api/src/log-analysis/file-validator.service.ts`  | `Express.Multer.File` | `void` ou erreur           |
 | `LoggerInterceptor`          | Journalisation globale des requêtes        | `apps/api/src/common/logger.interceptor.ts`      | `Request/Response` | `Observable`                |
 | `JsPdfGenerator`             | Génère et télécharge un rapport PDF        | `apps/web/src/lib/JsPdfGenerator.ts`            | `ParsedLog` | Fichier téléchargé |
+| `SortableTable`              | Tableau générique triable et filtrable     | `packages/ui-components/src/SortableTable.tsx` | `data`, `columns` | Composant React |
 
 ## 6. Détails par agent
 
@@ -110,6 +111,13 @@
 - **Sorties** : fichier PDF téléchargé.
 - **Dépendances** : `jsPDF`, `jspdf-autotable`.
 - **Tests** : `apps/web/src/__tests__/PdfButton.test.tsx` via injection.
+
+### `SortableTable`
+- **Rôle** : afficher un tableau triable avec filtrage global.
+- **Entrées** : `data`, `columns`, options de tri initial.
+- **Sorties** : rendu React.
+- **Dépendances** : `@tanstack/react-table`.
+- **Tests** : via `ErrorTable` dans `apps/web/src/__tests__/ErrorTable.test.tsx`.
 
 ## 7. Schéma d’interaction
